Add in old banner for testing purposes
When we test the modal on stage, we want the new banner that links to
the modal to show, and not the old static one that links to account
settings.

This places the modal version behind the "ENABLE_DEMOGRAPHICS_MODAL"
feature flag to be turned on in stage.

Also fixes the props for the token refresh