Reduce noise in CourseOverview->Modulestore fallback
This was originally intended as a placeholder to remind us to
optimize the data model for sequence metadata API calls (say
by adding those fields to the CourseOverview model itself).
Logging a quarter million messages a day is overkill for this
purpose, and we will be doing another pass at the performance of
this code in the coming weeks/months.