Use AppConfigs to ensure tasks discovery in bulk_email, ccxcon, bookmarks
This intends to further fix issues where some Celery tasks are
not being registered. We'd previously relied on auto-discovery
of many Celery tasks, but since the Celery 4 upgrade, it seems
that we need to explicitly ensure that `tasks.py` modules are
loaded by Celery workers.

We do this by:
* making sure an AppConfig subclass is declared in `apps.py` of each app,
* making sure it defines a `ready` method that imports `.tasks`, and
* making sure that INSTALLED_APPS points to the AppConfig subclass OR
  making sure that the app is installed as a Django app plugin via the
  AppConfig subclass.

TNL-7652