Don't install instructor_task for common testing.
We shouldn't need to access the instructor tasks and load them as a
part of the common testing.  Loading it here now that we load the
tasks.py module as a part of setting up this app causes a bunch of
downstream apps to be loaded.  None of which should be need for
common testing.