Don't install instructor_task for common testing. (#25459)
* Import instructor tasks explicitly on workers.

Importing them via app config leads to an dependency chain nightmare as lots of other apps get pulled into the startup
and are not configured in situations like testing the common and lib unit tests.  This should register tasks to the celery
workers without messing with the startup dependency chains of the LMS, CMS and test setup.