Run subsection completion milestones task async.
This function is a celery task but it seems like it is not being
called correctly.  The way it was, it would run in process. We
update the call so that it is called asynchronously as a celery
task instead.

This may have some user facing impact as the work might now take
longer to do than when it was running in the request process.

However leaving it as is, definitely makes the user experience slower
causes us to do computationally intensive work in the web request. It
also makes it much harder to monitor as our ownership assignment tracking
for the feature does not work correctly.