Fix import_shims pylint violations; lint import_shims
The import_shims/ directory had thousands of pylint
violations, but they didn't matter because we ignored
the directory during linting.

Now, that directory ignoring configuration seems to be
problematic when running pylint on individual files locally.

Also, it seems prudent to lint import_shims/ anyway,
because there is still production code that relies on it.