Remove all models related to shoppingcart.
DEPR-43 - Removal of Shoppingcart.

All models have been removed and the change has been deployed to production
so this should be safe to do.

It's strange that the ORM decided to drop certain fields before dropping the
whole models but decided to leave it all as the ORM generated it rather than
trying to manually change it.

Also turn the check back on to make sure people don't actually check in code
without the relevant migration files.