Pin requirements to temporarily fix instructor task registration
We ran into an issue where instructor tasks were not being registered with celery correctly, resulting in: https://openedx.atlassian.net/browse/CR-2982

The cause wasn’t clear, so we started reverting some recent, suspect PRs.  When we reverted #25746, the issue went away.

The revert PR was this one: https://github.com/edx/edx-platform/pull/25766

So that we can unpause our deployment pipelines, we’re temporarily pinning the versions of the four packages upgraded in #25746 so we can continue to investigate.