Downgrade REDIRECT_TO_COURSEWARE_MICROFRONTEND to CourseWaffleFlag
The toggle was previously an ExperimentWaffleFlag,
which allows stable A/B testing but increases the toggle's
complexity. Since we do not plan an doing any more A/B
tests as part of the MFE rollout, we can 'downgrade' this
toggle to a CourseWaffleFlag, which still allows us to do
phased rollout and course-run-specific overrides.

Also, link to DEPR-109 in courseware toggle annotations.

TNL-7505