Update sandbox requirement mechanism.
We seperated base and py35 as a part of the upgrade to python 3 so
that we could get newer versions of some requirements during that
upgrade.  This change is the first step in removing the old requirements
files that were only relevant to python 2.

We leave the python version centric named files instead of base because
we'll need to bump the version of this file regularly and clearly
indicating what version of python it's for makes it easier to reason
about.