Fix hiding of completion/bookmark indicators in legacy courseware (#25919)
In the edx.org-next theme, the completion green check-
marks and the little bookmark icons were appearing on every
single unit. This was because a fontawesome CSS rule
in the new theme was overriding the `display: none`,
as set by `.is-hidden`, to be `display: inline: block`.

The fix is to add `!important` to a couple definitions
of `.is-hidden`. Definitely somewhat hacky, but this is
SCSS that we plan to deprecate at some point anyway.

TNL-7823