Add user_tasks to LMS (for Django admin access).
The user_tasks app is used by Studio import/export and not used by
anything in the LMS. But because our app-permissions user/role
provisioning script invokes the LMS process, any app+permission
listed there must exist in the LMS or the provisioning fails. So
this commit adds the user_tasks app to the LMS list of installed
apps because it's a lot faster than getting app-permissions to do
the right thing.