Make DeprecatedEdxPlatformImportError derive ImportError
The latter was previously NOT a subclass of the
former. This was intentional, because I wanted to
make sure that code of the form:

try:
    from <deprecated_path> import X
except ImportError:
    X = None

failed *loudly* instead of silently.

Now that the import strictness has been on Stage
for a bit, and we have reasonably verified that all
such imports have been fixed, we must make
`DeprecatedEdxPlatformImportError` a subclass of
`ImportError`.

Not doing so is messing up
a signal handler in Django User Tasks that uses
`try: ... except ImportError: ...` as a way of
registering tasks, expecting to get ImportErrors
on tasks whose names are not importable. By
raising an incompatible error class, we are breaking
that signal hander.