edx-enterprise version 3.17.21
Introduce and use a roles_api module, use the roles API in receivers that want to create or delete role assignments.
For new learners and admins, associate their user-role with the customer to which they are linked.
Does not yet specify the enterprise_customer for deletions. We want to maintain current behavior on deletes until we do a data backfill of that field. So for now, we'll continue to delete all assignments held by the deleted user.
Install django-cache-memoize, so that we can memoize roles (not assignments) in the Django cache.