feat: Install version 3.1.0 of the edx_proctoring library into edx-platform.
This release adds a new Course Onboarding Status endpoint to the LMS. This endpoint returns information about learners' onboarding status, as indicated by edX systems. The statuses are meant to be externally facing and will be used on the Instructor Dashboard for the benefit of Course Authors, Operators, and Developers.