feat: bump version of edx-proctoring library to 3.6.2: change "proctoring_started" to "onboarding_started" in Student Onboarding Status panel, make time_remaining_seconds model field read only, and fix message bug on the Django admin panel for ProctoredExamStudentAttempt model
This release changes learner onboarding status from "proctoring_started" to "onboarding_started" to more clearly describe the learner's onboarding status in the Instructor Dashboard Student Onboarding Status panel. This impacts course authors, developers, and operators.

This release adds the time_remaining_seconds field of ProctoredExamStudentAttempt model to readonly_fields in Django admin page so it is not required when editing the model. This impacts developers.

This release updates reference to Exception.message to use string representation of the exception, as message is no longer an attribute of the Exception class. This impacts developers.