fix: course outlines should include problemsets & videosets
This commit fixes a utility function that is used in several
places (including the new courseware experience and both
the new and legacy course homes) to include problemsets
and videosets in its query to the block transformers cache.

problemset and videoset are both aliases to
sequential (e.g. Subsection). They are purely semantic;
they aren't supposed to affect functionality at all
currently.

TNL-7978