fix: Allow Discussions Configuration creation
There is currently an error on the /add page, due to a database
inconsistency when an null column exists where the schema says it
should not...

This appears to be blowing up the rendering of the LtiConfiguration
widget within the Discussions Configurations creation page, causing it,
too, to break.

Hopefully, by excluding invalid entries from the admin dropdown, we can
circumvent this issue (inside the LTI XBlock).